requirements.txt
click~=8.1.3
cloudscraper~=1.2.58
pyyaml~=6.0
rich~=12.5.1
tenacity~=8.0.1
python-dotenv==1.0.0